+++ oap-server/server-fetcher-plugin/kafka-fetcher-plugin/src/main/java/org/apache/skywalking/oap/server/analyzer/agent/kafka/KafkaFetcherHandlerRegister.java
@@ -81,7 +81,9 @@ public class KafkaFetcherHandlerRegister implements Runnable {
             config.getTopicNameOfMetrics(),
             config.getTopicNameOfProfiling(),
             config.getTopicNameOfTracingSegments(),
-            config.getTopicNameOfMeters()
+            config.getTopicNameOfMeters(),
+            config.getTopicNameOfLogs(),
+            config.getTopicNameOfJsonLogs()
         ))
                                               .values()
                                               .entrySet()